==> notebooks/2_fine_tuning_label.ipynb <==
import os
os.chdir('../')
import sys
import pandas as pd
import numpy as np

pd.set_option('display.max_rows', 80)
pd.set_option('display.max_colwidth', None)

# Load data

labeled_data_file = 'data/processed/labeled_output.csv'

df = pd.read_csv(labeled_data_file)

# The free models werent adequate because of API call limit
model_list = ['claude', 'gpt4']

df['label_combo'] = df[[f'rating_{m}' for m in model_list]].agg(tuple, axis=1)
df['label_combo'].value_counts().head(10)

df["agreement_score"] = df[[f'rating_{m}' for m in model_list]].nunique(axis=1)

df["agreement_score"].value_counts()

df['target'] = np.nan
df.loc[df['agreement_score']==1, 'target'] = df.loc[df['agreement_score']==1, 'rating_claude']

# Manual check consistency in the cases with no agreement

unagreement_index = df[df['agreement_score']==2].index

batch_size = 10
batches = [unagreement_index[i:i + batch_size] for i in range(0, len(unagreement_index), batch_size)]

df.loc[batches[0], ["text", "rating_claude", "rating_gpt4"]]

target = [-1, 0, 0, 0, -1, 0, 1, -1, 0, 1]
df.loc[batches[0], 'target'] = target

df.loc[batches[1], ["text", "rating_claude", "rating_gpt4"]]

target = [0, 0, -1, 0, -1, -1, 1, 1, -1, 0]
df.loc[batches[1], 'target'] = target

df.loc[batches[2], ["text", "rating_claude", "rating_gpt4"]]

target = [0, 1, -1, 0, 0, 1, 1, 0, -1, 1]
df.loc[batches[2], 'target'] = target

df.loc[batches[3], ["text", "rating_claude", "rating_gpt4"]]

target = [0, -1, 0, -1, 0, 0, -1, 0, -1, 0]
df.loc[batches[3], 'target'] = target

df.loc[batches[4], ["text", "rating_claude", "rating_gpt4"]]

target = [0, -1, -1, 1, 0]
df.loc[batches[4], 'target'] = target

# Check neutral 

df['target'].value_counts()

df.loc[df['target']==0, ['text', 'target']]

# Processing

df_final = df.loc[df['target']!=0, ['text', 'target']]

df_final['target'] = df_final['target'].replace(-1, 0)

df_final['target'].value_counts()

df_final.to_csv('data/processed/final_label.csv', index=False)
