==> notebooks/3_processing_step.ipynb <==
import os
os.chdir('../')

import sys
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

import re
import unicodedata
import string
from collections import Counter

from wordcloud import WordCloud

import nltk
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.sentiment import SentimentIntensityAnalyzer

from sklearn.feature_extraction.text import CountVectorizer

import spacy

from textblob import TextBlob

from symspellpy.symspellpy import SymSpell, Verbosity

import emoji

from src.llm_preprocessing import generate_whitelist

pd.set_option('display.max_rows', 30)
pd.set_option('display.max_colwidth', None)

# nltk.download("punkt")
# nltk.download("wordnet")
# nltk.download("stopwords")
# nltk.download('omw-1.4')
# nltk.download('vader_lexicon')

sym_spell = SymSpell(max_dictionary_edit_distance=2, prefix_length=7)
dictionary_path = 'data/dictionaries/frequency_dictionary_en_82_765.txt'  # included in the repo
sym_spell.load_dictionary(dictionary_path, term_index=0, count_index=1)

df_pretreatment = pd.read_csv('data/processed/final_label.csv')
df = df_pretreatment.copy()

df.head(5)

# Text processing pipeline

## Clean text

### Check for emojis

def contains_emoji(text):
    return any(char in emoji.EMOJI_DATA for char in text)

df['has_emoji'] = df['text'].apply(contains_emoji)

# See how many have emojis
print(df['has_emoji'].value_counts())

There are no emojis, but it is important to keep in the pipeline in case the test set has

### Check for numbers

df['has_number'] = df['text'].str.contains(r"\d")
print(df['has_number'].value_counts())

df[df['has_number']].head(3)

df['has_rating_number'] = df['text'].str.contains(r"\b(?:give it a|rated?|score|rate|i give|i give this|its a|it's a|i rate|rating|^)(?:\s+)?(?:10|[0-9])\b", flags=re.IGNORECASE)

df[df['has_rating_number']]

Most numbers are useless but some cases contain ratings. The feature has_rating_number captures this using a common words associated with a rate

### Clean text function

def clean_text(text):
    text = text.lower() # Lower case
    text = emoji.demojize(text) # transform emojis
    text = re.sub(r"\d+", "", text)  # remove digits
    text = re.sub(r"[^\w\s!]", "", text)  # keep only words, spaces, and !
    text = re.sub(r"\s+", " ", text).strip() # remove whitespaces
    return text

df['text_v1'] = df['text'].apply(clean_text)

## Remove typos

The idea is to correct some typos using sym_spell dictionary. Some food names such as mizu, confit, foie gras, etc can be inadequately replaced. To evaluate with that a new column is created with all corrections made per text

correction_counter = Counter()
blocklist = {'the': 'the', '!': '!', 's': 's', 'n': 'n', 'wa': 'was'}

def correct_typos(text, blocklist={'the': 'the', '!': '!', 's': 's', 'n': 'n', 'wa': 'was'}, whitelist=[]):
    corrected_words = []
    corrections_in_row = []
    
    for word in text.split():
        # Always respect blocklist and whitelist first
        if word in blocklist or word in whitelist or len(word) <= 2:
            corrected_words.append(word)
            continue
        
        # Try correcting
        suggestions = sym_spell.lookup(word, Verbosity.CLOSEST, max_edit_distance=2)
        
        if suggestions and suggestions[0].term != word:
            correction_counter[(word, suggestions[0].term)] += 1
            corrected_words.append(suggestions[0].term)
            corrections_in_row.append(word)  # Log the original word
        else:
            corrected_words.append(word)
    
    return " ".join(corrected_words), "; ".join(corrections_in_row) if corrections_in_row else None


df[['corrected_text', 'corrections']] = df['text_v1'].apply(lambda x: pd.Series(correct_typos(x)))

df[['corrected_text', 'corrections']].loc[~df['corrections'].isna()].head()

### Creating a custom whitelist using llm

With the corrections column, the idea is to create a whitelist to avoid wrong corrections. To do that, a gpt4 model is called to evaluate each of the corrections and say if they are adequate or not. Using those results the whitelist is created. It is then hardcoded to avoid having to run the llm everytime

df_corrections = df.loc[df['corrections'].notna(), ["corrected_text", "corrections"]]
# results = generate_whitelist(df_corrections, 'corrected_text', model_name='gpt4')
# df_corrections['result'] = results

def apply_whitelist(df):
    whitelist_set = set()

    for idx, row in df.iterrows():
        words = [w.strip() for w in row['corrections'].split(';') if w.strip()]
        results = row['result']

        for word, flag in zip(words, results):
            if flag == 1:
                whitelist_set.add(word)

    return sorted(whitelist_set)

# whitelist = apply_whitelist(df_corrections)

whitelist = ['branzino', 'carozza', 'confit', 'congee', 'cozy', 'dal', 'facie', 'favorite', 'fixe', 'flavoring', 'foie', 'frites', 'gras', 'grazie', 'gulab', 'jamun', 'kha', 'maitre', 'medicore', 'mizu', 'mom', 'msg', 'neighborhood', 'noir', 'nyc', 'overpack', 'porcini', 'prima', 'prixe', 'russe', 'shabu', 'svc', 'tartare', 'theater', 'tristate', 'uncourteous', 'volare', 'wks']

### Final typo corrections

With the withelist a new column with fine-processed text is created

df['text_v2'] = df["text_v1"].apply(lambda x: pd.Series(correct_typos(x, blocklist, whitelist)))[0]

df.drop(columns=['corrected_text', 'corrections'], inplace=True)

## Lemmatize and stopwords

Lemmatize and stopwords will be evaluated in more details in the EDA phase. Spacy was used instead of nltk because it was lemmatizing 'us' to 'u'. Some custom stopwords were added

nlp = spacy.load('en_core_web_sm')

def spacy_lemmatize(text):
    doc = nlp(text)
    return " ".join([token.lemma_ for token in doc])

stop_words = set(stopwords.words("english"))
custom_stopwords = {'I', 'i', 'n', 'wa', 's'}
stop_words.update(custom_stopwords)

def remove_stopwords(text):
    tokens = text.split()
    return " ".join([w for w in tokens if w not in stop_words])

# Full treatment pipeline

Incorporating all steps in a pipeline

def full_preprocess(text):
    text = clean_text(text)
    text = correct_typos(text, blocklist, whitelist)[0]
    text = spacy_lemmatize(text)
    text = remove_stopwords(text)
    return text

df

df = df_pretreatment.copy()

df = df_pretreatment.copy()
df['has_number'] = df['text'].str.contains(r"\d")
df['has_rating_number'] = df["text"].str.contains(
    r"\b(?:give it a|rated?|score|rate|i give|i give this|its a|it's a|i rate|rating|^)(?:\s+)?(?:10|[0-9])\b",
    flags=re.IGNORECASE
)
df['final_text'] = df['text'].apply(full_preprocess)

def light_clean_text_for_bert(text):
    text = emoji.demojize(text)
    text = re.sub(r"\s+", " ", text).strip()
    return text

df['bert_text'] = df['text'].apply(light_clean_text_for_bert)

df.info()

df.to_csv('data/processed/preprocessed_train.csv', index=False)
